breaks out cache functions to their own file
